refactor: change session's co-routine with stateful class, ref #17
