+++ code/ac/llama/Session.hpp
@@ -11,207 +11,41 @@
 #include <cassert>
 
 namespace ac::llama {
+class Instance;
 
 class Session {
 public:
-    struct Prompt {}; // sentinel for co_await-int prompts
-    struct StartGeneration {}; // sentinel for co_await-ing the start of generation
+    struct InitParams {
+        uint32_t gaFactor = 1; // group-attention factor
+        uint32_t gaWidth = 512; // group-attention width
 
-    struct SessionOpData {
-        enum OpType {
-            Prompt,
-            GetState,
-            SetState,
-
-            Count
-        };
-        OpType type = Count;
-
-        std::span<const Token> pendingPrompt;
-        std::span<uint8_t> state;
+        // if true, the inference tries to extend the context by truncating previous tokens
+        // only used if gaFactor == 1
+        bool infiniteContext = true;
     };
+    Session(Instance& instance, InitParams params);
 
-    struct SessionResult {
-        enum class Type {
-            Token,
-            State,
-            Bool,
-            Invalid
-        };
+    void setInitialPrompt(std::span<const Token> prompt);
 
-        Type type = Type::Invalid;
-        Token token = Token_Invalid;
-        std::vector<uint8_t> state;
-        bool result = false;
-    };
-
-    class promise_type {
-        SessionResult m_value;
-
-        SessionOpData m_pendingOpData;
-
-        std::exception_ptr m_exception;
-    public:
-        Session get_return_object() noexcept {
-            return Session{std::coroutine_handle<promise_type>::from_promise(*this)};
-        }
-
-        SessionResult value() const noexcept { return m_value; }
-
-        // suspend until the initial prompt is set
-        std::suspend_always initial_suspend() noexcept { return {}; }
-
-        // keep the coroutine alive to make the last yield value available
-        std::suspend_always final_suspend() noexcept { return {}; }
-
-        std::suspend_always yield_value(Token value) noexcept {
-            m_value.type = SessionResult::Type::Token;
-            m_value.token = value;
-            return {};
-        }
-
-        std::suspend_always yield_value(std::vector<uint8_t>& stateData) noexcept {
-            m_value.type = SessionResult::Type::State;
-            m_value.state = std::move(stateData);
-            return {};
-        }
-
-        std::suspend_always yield_value(bool result) noexcept {
-            m_value.type = SessionResult::Type::Bool;
-            m_value.result = result;
-            return {};
-        }
-
-        void return_void() noexcept {
-            m_value.type = SessionResult::Type::Invalid;
-        }
-        void unhandled_exception() noexcept {
-            // why store the exception here instead of simply rethrowing?
-            // because clang is stupid, that's why
-            // clang's ridiculous handling of coroutines causes local coroutine variables to be destroyed twice if we
-            // just rethrow here
-            m_exception = std::current_exception();
-        }
-
-        struct Awaiter {
-            promise_type& self;
-            bool await_ready() noexcept { return true; }
-            SessionOpData await_resume() noexcept {
-                // clear pending after returning it
-                return std::exchange(self.m_pendingOpData, SessionOpData());
-            }
-            void await_suspend(std::coroutine_handle<>) noexcept {}
-        };
-
-        void setOpType(SessionOpData::OpType type) {
-            m_pendingOpData.type = type;
-        }
-
-        void setPrompt(std::span<const Token> prompt) {
-            assert(m_pendingOpData.type == SessionOpData::Count);
-            m_pendingOpData.type = SessionOpData::Prompt;
-            m_pendingOpData.pendingPrompt = prompt;
-        }
-
-        void setState(std::span<uint8_t> state) {
-            assert(m_pendingOpData.type == SessionOpData::Count);
-            m_pendingOpData.type = SessionOpData::SetState;
-            m_pendingOpData.state = state;
-        }
-
-        void getState() {
-            assert(m_pendingOpData.type == SessionOpData::Count);
-            m_pendingOpData.type = SessionOpData::GetState;
-        }
-
-        Awaiter await_transform(Prompt) noexcept { return Awaiter{*this}; }
-
-        std::suspend_always await_transform(StartGeneration) noexcept { return {}; }
-
-        void rethrowIfException() {
-            if (m_exception) {
-                std::rethrow_exception(m_exception);
-            }
-        }
-    };
-
-    using Handle = std::coroutine_handle<promise_type>;
-
-    Session() = default;
-    Session(Handle handle) : m_handle(handle) {}
-
-    Session(Session&& other) noexcept : m_handle(std::exchange(other.m_handle, nullptr)) {}
-    Session& operator=(Session&& other) noexcept {
-        if (this == &other) return *this;
-        if (m_handle) {
-            m_handle.destroy();
-        }
-        m_handle = std::exchange(other.m_handle, nullptr);
-        return *this;
-    }
-
-    ~Session() {
-        if (m_handle) {
-            m_handle.destroy();
-        }
-    }
-
-    void reset() {
-        if (m_handle) {
-            m_handle.destroy();
-        }
-        m_handle = nullptr;
-    }
-
-    explicit operator bool() const noexcept { return !!m_handle; }
-
-    // the provided span must remain valid until the next call to getToken or pushPrompt with another span
-    void pushPrompt(std::span<const Token> prompt) {
-        m_handle.promise().setPrompt(prompt);
-    }
-
-    void setInitialPrompt(std::span<const Token> prompt) {
-        m_handle.promise().setPrompt(prompt);
-        m_handle.resume();
-        m_handle.promise().rethrowIfException();
-    }
-
-    Token getToken() {
-        if (m_handle.done()) return Token_Invalid;
-        m_handle.promise().setOpType(SessionOpData::Prompt);
-        m_handle.resume();
-        m_handle.promise().rethrowIfException();
-
-        assert(m_handle.promise().value().type == SessionResult::Type::Token);
-        m_handle.promise().setOpType(SessionOpData::Count);
-
-        return std::move(m_handle.promise().value().token);
-    }
-
-    std::vector<uint8_t> getState() {
-        m_handle.promise().getState();
-        m_handle.resume();
-        m_handle.promise().rethrowIfException();
-
-        assert(m_handle.promise().value().type == SessionResult::Type::State);
-        m_handle.promise().setOpType(SessionOpData::Count);
-
-        return std::move(m_handle.promise().value().state);
-    }
-
-    bool setState(std::span<uint8_t> state) {
-        m_handle.promise().setState(state);
-        m_handle.resume();
-        m_handle.promise().rethrowIfException();
-
-        assert(m_handle.promise().value().type == SessionResult::Type::Bool);
-        m_handle.promise().setOpType(SessionOpData::Count);
-
-        return std::move(m_handle.promise().value().result);
-    }
-
+    void pushPrompt(std::span<const Token> prompt);
+    Token getToken();
+    std::vector<uint8_t> getState();
+    bool setState(std::span<uint8_t> state);
 private:
-    Handle m_handle;
+    enum class Source {
+        InitialPrompt,
+        InteractivePrompt,
+        Generated
+    };
+
+    void doDecode(std::span<const Token> tokens, Source src);
+
+    Instance& m_instance;
+    InitParams m_params;
+    unsigned maxTokens = 0;
+    unsigned numKeep = 0;
+    uint32_t gaIndex = 0; // number of grouped KV tokens (only used if params.gaFactor > 1)
+    uint32_t numPast = 0; // number of tokens in the context (that's prompts + generated)
 };
 
 } // namespace ac::llama
